fix(sync): duration overflow, LWW merge defaults, API double-encode, DTO type fixes
Phase 6 sync-data cluster: use Long for PortalSyncAdapter duration to prevent Int
overflow on long sessions, fix SqlDelightSyncRepository mergeSessionsLww timestamp
default and GamificationStats Float→Long truncation, fix mergePortalCycles defaults,
prevent PortalApiClient double URL-encoding, update SyncModels duration type to Long.

+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/data/sync/SyncModels.kt
@@ -41,7 +41,8 @@ data class WorkoutSessionSyncDto(
     val mode: String,
     val targetReps: Int,
     val weightPerCableKg: Float,
-    val duration: Int = 0,
+    /** Session duration in milliseconds. Matches [com.devil.phoenixproject.domain.model.WorkoutSession.duration]. */
+    val duration: Long = 0L,
     val totalReps: Int = 0,
     val exerciseId: String? = null,
     val exerciseName: String? = null,
